+++ src/binary16.rs
@@ -188,7 +188,7 @@ impl f16 {
     /// assert!(!f.is_nan());
     /// ```
     #[inline]
-    pub fn is_nan(self) -> bool {
+    pub const fn is_nan(self) -> bool {
         self.0 & 0x7FFFu16 > 0x7C00u16
     }
 
@@ -212,7 +212,7 @@ impl f16 {
     /// assert!(neg_inf.is_infinite());
     /// ```
     #[inline]
-    pub fn is_infinite(self) -> bool {
+    pub const fn is_infinite(self) -> bool {
         self.0 & 0x7FFFu16 == 0x7C00u16
     }
 
@@ -235,7 +235,7 @@ impl f16 {
     /// assert!(!neg_inf.is_finite());
     /// ```
     #[inline]
-    pub fn is_finite(self) -> bool {
+    pub const fn is_finite(self) -> bool {
         self.0 & 0x7C00u16 != 0x7C00u16
     }
 
@@ -341,7 +341,7 @@ impl f16 {
     /// assert!(nan.is_sign_positive() != nan.is_sign_negative());
     /// ```
     #[inline]
-    pub fn is_sign_positive(self) -> bool {
+    pub const fn is_sign_positive(self) -> bool {
         self.0 & 0x8000u16 == 0
     }
 
@@ -363,7 +363,7 @@ impl f16 {
     /// assert!(nan.is_sign_positive() != nan.is_sign_negative());
     /// ```
     #[inline]
-    pub fn is_sign_negative(self) -> bool {
+    pub const fn is_sign_negative(self) -> bool {
         self.0 & 0x8000u16 != 0
     }
 }