+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Added `bfloat16` type for alternative 16-bit floating point conversions. By [@tspiteri].
-- `f16::from_bits` and `f16::to_bits` are now `const` fns.
+- `f16::from_bits`, `f16::to_bits`, `f16::is_nan`, `f16::is_infinite`, `f16::is_finite`,
+  `f16::is_sign_positive`, and `f16::is_sign_negative` are now `const` fns.
 - `slice::HalfBitsSliceExt` and `slice::HalfBitsSliceExt` extension traits have been added for
   performing efficient reinterpret casts and conversions of slices to and from `[f16]` and
   `[bf16]`.  These traits will use hardware SIMD conversion instructions when available and the